ApiRelay/requirements.txt
fastapi==0.53.2
pymongo==3.10.1
python-dotenv==0.12.0
requests==2.23.0
uvicorn==0.11.3
xmltodict==0.12.0
